exchange, share, ...

+++ Service/Service/ShareDetailService.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
+using VG.Common.Constant;
 using VG.Data.Repository;
 using VG.Model.Model;
 using VG.Model.RequestModel;
@@ -15,6 +16,7 @@ namespace VG.Service.Service
     {
         ShareDetail Add(ShareDetailRequestModel newItem, string phoneNumber);
         IEnumerable<ShareDetailResponseModel> GetShareByAccountId(string phoneNumber);
+        IEnumerable<ShareDetailResponseModel> GetAll (string phoneNumber);
         ShareDetailResponseModel Get(string Id);
         ShareDetailRequestModel Update(ShareDetailRequestModel newItem, string phoneNumber, string savePath, string domain);
         void Delete(string Id);
@@ -25,12 +27,16 @@ namespace VG.Service.Service
         private IAccountRepository _accountRepository;
         private IVegetableService _vegetableService;
         private IAccountDetailRepository _accountDetailRepository;
-        public ShareDetailService(IShareDetailRepository shareRepository, IAccountRepository accountRepository, IVegetableService vegetableService, IAccountDetailRepository accountDetailRepository)
+        private IExchangeDetailRepository _exchangeDetailRepository;
+        private IAccountFriendRepository _accountFriendRepository;
+        public ShareDetailService(IShareDetailRepository shareRepository, IAccountRepository accountRepository, IVegetableService vegetableService, IAccountDetailRepository accountDetailRepository, IExchangeDetailRepository exchangeDetailRepository, IAccountFriendRepository accountFriendRepository)
         {
             _shareRepository = shareRepository;
             _accountRepository = accountRepository;
             _vegetableService = vegetableService;
             _accountDetailRepository = accountDetailRepository;
+            _exchangeDetailRepository = exchangeDetailRepository;
+            _accountFriendRepository = accountFriendRepository;
         }
         public ShareDetail Add(ShareDetailRequestModel newItem, string phoneNumber)
         {
@@ -43,6 +49,8 @@ namespace VG.Service.Service
                 Status = newItem.Status,
                 Quantity = newItem.Quantity,
                 VegetableId = newItem.VegetableId,
+                VegetableNeedId = newItem.VegetableNeedId,
+                Lock = false
             });
             this._shareRepository.Commit();
             return share;
@@ -50,7 +58,11 @@ namespace VG.Service.Service
 
         public void Delete(string Id)
         {
-            var share = this._shareRepository.GetSingle(s => s.Id == Id);
+            var share = this._shareRepository.GetSingle(s => s.Id == Id, new string[] { "ExchangeDetails" });
+            if (share.ExchangeDetails.Where(s => s.Status == (int)EnumStatusRequest.Shipping || s.Status == (int)EnumStatusRequest.Accept).Count() > 0)
+            {
+                throw new Exception("Có đơn hàng đang được vận chuyển, không thể xoá bài viết này");
+            }
             this._shareRepository.Delete(share);
             this._shareRepository.Commit();
         }
@@ -74,6 +86,18 @@ namespace VG.Service.Service
                 Statius = result.Status,
                 Images = veg.Images
             };
+        }
+
+        public IEnumerable<ShareDetailResponseModel> GetAll(string phoneNumber)
+        {
+            var account = this._accountRepository.GetSingle(s => s.PhoneNumber == phoneNumber);
+            var listFriend = this._accountFriendRepository.GetMulti(s => s.Account_one_Id == account.Id).Select(s => s.Account_two_Id).ToList();
+            var listFriend2 = this._accountFriendRepository.GetMulti(s => s.Account_two_Id == account.Id).Select(s => s.Account_one_Id).ToList();
+            listFriend.AddRange(listFriend2);
+            var shareFriend = this._shareRepository.GetAll(listFriend).ToList();
+            var allShare = this._shareRepository.GetAll().Where(s => !shareFriend.Select(q => q.Id).Contains(s.Id)).ToList();
+            var result = shareFriend.Union(allShare);
+            return result;
         }
 
         public IEnumerable<ShareDetailResponseModel> GetShareByAccountId(string phoneNumber)
